+++ R/launch_shiny.R
@@ -1,6 +1,6 @@
 # launch shiny app
 
-launchShiny <- function(shiny_app) {
+launch_shiny <- function(shiny_app) {
   appDir <- system.file("shiny", shiny_app, package = "qlcData")
   shiny::runApp(appDir, display.mode = "normal")
 }