+++ R/asPhylo.R
@@ -1,17 +1,17 @@
 asPhylo <- function(tree, height = 100, fixed.branches = NULL, long.root = NULL) {
 	
   # this export to newick takes time with large trees
-  tree <- data.tree::FromDataFrameNetwork(tree)
-	newick <- data.tree::ToNewick(tree, rootHeight = height)
-	phylo <- phytools::read.newick(text = newick)
+  # tree <- data.tree::FromDataFrameNetwork(tree)
+	# newick <- data.tree::ToNewick(tree, rootHeight = height)
+	# phylo <- phytools::read.newick(text = newick)
 	
   # this simple function cannot take 'height' and always takes 'fixed = 1'
-  # phylo <- convertTreePhylo(tree)
+  phylo <- convertTreePhylo(tree, height =  height/2)
   
 	if (!is.null(fixed.branches)) {
-		phylo$edge.length <- rep(fixed.branches, length(phylo$edge.length))
-	}
-	
+		phylo$edge.length <- rep(fixed.branches, length = nrow(phylo$edge))
+	} 
+  
 	if (!is.null(long.root)) {
 		roots <- c("World"
 		          , "Southeast_Asia"
@@ -30,8 +30,8 @@ asPhylo <- function(tree, height = 100, fixed.branches = NULL, long.root = NULL)
 	phylo <- ape::multi2di(ape::collapse.singles(phylo))
 
 	# Spaces were removed by newick conversion. Reintroduced here.
-	phylo$node.label <- gsub("_", " ", phylo$node.label)
-	phylo$tip.label <- gsub("_", " ", phylo$tip.label)
+	# phylo$node.label <- gsub("_", " ", phylo$node.label)
+	# phylo$tip.label <- gsub("_", " ", phylo$tip.label)
 	
 	return(phylo)
 }