+++ NEWS.md
@@ -1,7 +1,7 @@
 # qlcData 0.2
 
-* including a shiny app for tokenization, with a helper launcher 'launchShiny()'
-* adding docopt-executables to \exec with a helper function to make a bash alias 'add_execs()'
+* including a shiny app for tokenization, with a helper launcher 'launch_shiny()'
+* adding docopt-executables to \exec with a helper function to make a bash alias 'add_alias()'
 * make sure orthography profiles are treated as character data, not as factors
 * adding function 'pass_align()' to transfer alignment to a string, with helper 'join_align()'
 * changed vignette naming