+++ NEWS.md
@@ -1,11 +1,11 @@
 # qlcData 0.2
 
-* add glottolog data with access function getTree()
+* add glottolog data with access function getTree() and asPhylo()
 * add possibility for factors in recoding
 * including a shiny app for tokenization, with a helper launcher 'launch_shiny()'
 * adding docopt-executables to \exec with some help how to softlink these to PATH
 * make sure orthography profiles are treated as character data, not as factors
-* adding function 'pass_align()' to transfer alignment to a string, with helper 'join_align()'
+* adding function pass_align() to transfer alignment to a string, with helper join_align()
 * changed vignette naming
 
 # qlcData 0.1